requirements.txt
click==8.0.2
djangorestframework==3.13.1
